set market mode upgrade mango lib (#780)
* fix

* fix
