+++ Strategies/protocols/mango/tools.ts
@@ -314,3 +314,41 @@ export const tryGetMangoAccountsForOwner = async (
     return null
   }
 }
+
+export const ASSET_TYPE = [
+  {
+    name: 'Token',
+    value: 0,
+  },
+  {
+    name: 'Perp',
+    value: 1,
+  },
+]
+
+export const MARKET_MODE = [
+  {
+    name: 'Default',
+    value: 0,
+  },
+  {
+    name: 'Active',
+    value: 1,
+  },
+  {
+    name: 'Close Only',
+    value: 2,
+  },
+  {
+    name: 'Force Close Only',
+    value: 3,
+  },
+  {
+    name: 'Inactive',
+    value: 4,
+  },
+  {
+    name: 'Swapping Spot Market',
+    value: 5,
+  },
+]
