+++ package.json
@@ -23,7 +23,7 @@
     ]
   },
   "dependencies": {
-    "@blockworks-foundation/mango-client": "^3.6.3",
+    "@blockworks-foundation/mango-client": "^3.6.5",
     "@blockworks-foundation/voter-stake-registry-client": "^0.2.3",
     "@cardinal/namespaces-components": "^2.5.5",
     "@castlefinance/vault-core": "^0.1.3",
@@ -37,8 +37,8 @@
     "@friktion-labs/friktion-sdk": "^1.1.118",
     "@headlessui/react": "^1.6.4",
     "@heroicons/react": "^1.0.1",
-    "@identity.com/solana-gateway-ts": "^0.8.2",
     "@hookform/resolvers": "^2.8.10",
+    "@identity.com/solana-gateway-ts": "^0.8.2",
     "@marinade.finance/marinade-ts-sdk": "^2.0.9",
     "@metaplex-foundation/mpl-token-metadata": "^1.2.5",
     "@mithraic-labs/serum-remote": "^0.0.1-rc.16",
